requirements.txt
google-generativeai
rapidfuzz
pillow
